Change circle points logic
Circle points now have more options and their logic changed.
Instead of building full circle we build only arc of this circle
So new options:
    - coordintaes: array of X and Y
        Represents the start of the arc
    - radius
        Represents radius of the arc
    - isConvex
        Tells the program reverse the arc of leave it like it is

So if coordinates option represent the start of the arc then the next point in the shape points array will
represent the end of arc.

There is 1 rule. The distance between start and end point cannot be greater than diameter